Move Property to a separate file
