+++ projects/com.oracle.truffle.llvm.test/src/com/oracle/truffle/llvm/test/TestHelper.java
@@ -40,6 +40,7 @@ import java.util.Set;
 import java.util.stream.Collectors;
 
 import com.oracle.truffle.llvm.runtime.options.LLVMOptions;
+import com.oracle.truffle.llvm.runtime.options.Property;
 import com.oracle.truffle.llvm.tools.Clang;
 import com.oracle.truffle.llvm.tools.Clang.ClangOptions;
 import com.oracle.truffle.llvm.tools.GCC;
@@ -165,8 +166,8 @@ public class TestHelper {
             String classpath = System.getProperty("java.class.path");
             String className = RemoteLLVMTester.class.getCanonicalName();
             String bootClassPath = LLVMOptions.getRemoteTestBootClassPath();
-            String debugOption = asOption(LLVMOptions.Property.DEBUG.getKey(), "false");
-            String projectRootOption = asOption(LLVMOptions.Property.PROJECT_ROOT.getKey(), LLVMOptions.getProjectRoot());
+            String debugOption = asOption(Property.DEBUG.getKey(), "false");
+            String projectRootOption = asOption(Property.PROJECT_ROOT.getKey(), LLVMOptions.getProjectRoot());
             String options = debugOption + " " + projectRootOption;
             String command = javaBin + " -cp " + classpath + " " + bootClassPath + " " + options + " " + className;
             Process process = Runtime.getRuntime().exec(command);
